refactor: standardize package naming, qualifiers, and variable language
- Rename uisystem package from `com.bill.minhadispensa` to `com.bitlabbr.minhadespensa` to match other modules
- Update related files and references to ensure consistent package naming across the project
- Replace string-based Koin qualifiers with type-safe qualifiers/constants
- Rename Portuguese variable `produtos` to an English equivalent for consistency and readability
