+++ composeApp/src/androidMain/AndroidManifest.xml
@@ -2,7 +2,7 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
 
     <application
-            android:name=".MinhaDespensaApplication"
+            android:name="com.bitlabbr.minhadespensa.app.MinhaDespensaApplication"
             android:allowBackup="true"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
@@ -11,7 +11,7 @@
             android:theme="@android:style/Theme.Material.Light.NoActionBar">
         <activity
                 android:exported="true"
-                android:name=".MainActivity">
+                android:name="com.bitlabbr.minhadespensa.app.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
