--- composeApp/src/androidMain/kotlin/com/bill/minhadispensa/minhadispensa/MainActivity.kt
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2026 Willian Santos
- *
- * Licensed under the Creative Commons Attribution-NonCommercial 4.0
- * International License (CC BY-NC 4.0).
- *
- * You may use, copy, modify, and distribute this file for non-commercial
- * purposes only, provided that proper attribution is given.
- *
- * The copyright holder retains all commercial rights and may
- * license this work under different terms.
- *
- * License: https://creativecommons.org/licenses/by-nc/4.0/
- */
-
-
-package com.bill.minhadispensa.minhadispensa
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.activity.enableEdgeToEdge
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.tooling.preview.Preview
-
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        enableEdgeToEdge()
-        super.onCreate(savedInstanceState)
-
-        setContent {
-            App()
-        }
-    }
-}
-
-@Preview
-@Composable
-fun AppAndroidPreview() {
-    App()
-}
