+++ composeApp/src/commonMain/kotlin/com/bitlabbr/minhadespensa/app/data/repository/FakeProductRepository.kt
@@ -1,0 +1,77 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * Licensed under the Creative Commons Attribution-NonCommercial 4.0
+ * International License (CC BY-NC 4.0).
+ *
+ * You may use, copy, modify, and distribute this file for non-commercial
+ * purposes only, provided that proper attribution is given.
+ *
+ * The copyright holder retains all commercial rights and may
+ * license this work under different terms.
+ *
+ * License: https://creativecommons.org/licenses/by-nc/4.0/
+ *
+ */
+
+package com.bitlabbr.minhadespensa.app.data.repository
+
+import com.bitlabbr.minhadespensa.core.domain.model.MeasureUnit
+import com.bitlabbr.minhadespensa.core.domain.model.Product
+import com.bitlabbr.minhadespensa.core.domain.repository.ProductRepository
+import com.bitlabbr.minhadespensa.core.domain.util.AppLogger
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import kotlinx.datetime.Clock
+
+class FakeProductRepository(
+    private val logger: AppLogger
+) : ProductRepository {
+
+    private val TAG = "FakeProductRepository"
+    private val bulkInsert = mutableListOf(
+        Product(
+            id = "1",
+            name = "Arroz Branco",
+            amount = 2.0,
+            unitMeasure = MeasureUnit.KILOGRAM,
+            expirationDate = Clock.System.now(),
+            imgUrl = null
+        ),
+        Product(
+            id = "2",
+            name = "Leite Integral",
+            amount = 4.0,
+            unitMeasure = MeasureUnit.LITER,
+            expirationDate = null,
+            imgUrl = null
+        ),
+        Product(
+            id = "3",
+            name = "Macarrão",
+            amount = 3.0,
+            unitMeasure = MeasureUnit.PACKAGE,
+            expirationDate = Clock.System.now(),
+            imgUrl = null
+        )
+    )
+
+    override fun getAllProducts(): Flow<List<Product>> = flow {
+        logger.d(TAG, "getAllProducts()")
+        emit(bulkInsert)
+    }
+
+    override suspend fun insertProduct(product: Product) {
+        bulkInsert.add(product)
+        logger.d(TAG, "Adding: ${product.name}")
+    }
+
+    override suspend fun deleteProductById(id: String) {
+        bulkInsert.remove(bulkInsert.find { it.id == id })
+        logger.d(TAG, "Removing ID: $id")
+    }
+
+    override suspend fun getProductById(id: String): Product? {
+        return bulkInsert.find { it.id == id }
+    }
+}
