--- composeApp/src/commonTest/kotlin/com/bill/minhadispensa/minhadispensa/ComposeAppCommonTest.kt
@@ -1,12 +1,0 @@
-package com.bill.minhadispensa.minhadispensa
-
-import kotlin.test.Test
-import kotlin.test.assertEquals
-
-class ComposeAppCommonTest {
-
-    @Test
-    fun example() {
-        assertEquals(3, 1 + 2)
-    }
-}
