+++ composeApp/src/iosMain/kotlin/com/bitlabbr/minhadespensa/app/Platform.ios.kt
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * Licensed under the Creative Commons Attribution-NonCommercial 4.0
+ * International License (CC BY-NC 4.0).
+ *
+ * You may use, copy, modify, and distribute this file for non-commercial
+ * purposes only, provided that proper attribution is given.
+ *
+ * The copyright holder retains all commercial rights and may
+ * license this work under different terms.
+ *
+ * License: https://creativecommons.org/licenses/by-nc/4.0/
+ *
+ */
+
+package com.bitlabbr.minhadespensa.app
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform : Platform {
+    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
