+++ core/src/commonMain/kotlin/com/bitlabbr/minhadespensa/core/domain/repository/ProductRepository.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * Licensed under the Creative Commons Attribution-NonCommercial 4.0
+ * International License (CC BY-NC 4.0).
+ *
+ * You may use, copy, modify, and distribute this file for non-commercial
+ * purposes only, provided that proper attribution is given.
+ *
+ * The copyright holder retains all commercial rights and may
+ * license this work under different terms.
+ *
+ * License: https://creativecommons.org/licenses/by-nc/4.0/
+ *
+ */
+
+package com.bitlabbr.minhadespensa.core.domain.repository
+
+import com.bitlabbr.minhadespensa.core.domain.model.Product
+import kotlinx.coroutines.flow.Flow
+
+interface ProductRepository {
+    fun getAllProducts(): Flow<List<Product>>
+    suspend fun insertProduct(product: Product)
+    suspend fun deleteProductById(id: String)
+    suspend fun getProductById(id: String): Product?
+}
