+++ uisystem/src/commonMain/kotlin/com/bitlabbr/minhadespensa/uisystem/di/DiQualifiers.kt
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * Licensed under the Creative Commons Attribution-NonCommercial 4.0
+ * International License (CC BY-NC 4.0).
+ *
+ * You may use, copy, modify, and distribute this file for non-commercial
+ * purposes only, provided that proper attribution is given.
+ *
+ * The copyright holder retains all commercial rights and may
+ * license this work under different terms.
+ *
+ * License: https://creativecommons.org/licenses/by-nc/4.0/
+ *
+ */
+
+package com.bitlabbr.minhadespensa.uisystem.di
+
+object DiQualifiers {
+    const val APP_LOGGER = "app_logger"
+    const val CORE_LOGGER = "core_logger"
+    const val UI_LOGGER = "ui_logger"
+}
