+++ uisystem/src/commonMain/kotlin/com/bitlabbr/minhadespensa/uisystem/di/UiModule.kt
@@ -1,0 +1,36 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * Licensed under the Creative Commons Attribution-NonCommercial 4.0
+ * International License (CC BY-NC 4.0).
+ *
+ * You may use, copy, modify, and distribute this file for non-commercial
+ * purposes only, provided that proper attribution is given.
+ *
+ * The copyright holder retains all commercial rights and may
+ * license this work under different terms.
+ *
+ * License: https://creativecommons.org/licenses/by-nc/4.0/
+ *
+ */
+
+package com.bitlabbr.minhadespensa.uisystem.di
+
+import com.bitlabbr.minhadespensa.core.domain.util.AppLogger
+import com.bitlabbr.minhadespensa.core.domain.util.ConsoleLogger
+import com.bitlabbr.minhadespensa.uisystem.features.list.ProductsListViewModel
+import org.koin.core.module.dsl.viewModel
+import org.koin.core.qualifier.named
+import org.koin.dsl.module
+
+val uiModule = module {
+    factory<AppLogger>(named(DiQualifiers.UI_LOGGER)) {
+        ConsoleLogger(moduleName = "UISystem")
+    }
+    viewModel {
+        ProductsListViewModel(
+            repository = get(),
+            logger = get(named(DiQualifiers.UI_LOGGER))
+        )
+    }
+}
