+++ uisystem/src/commonMain/kotlin/com/bitlabbr/minhadespensa/uisystem/theme/Color.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * This file is part of an open-source project licensed under the
+ * Creative Commons Attribution-NonCommercial 4.0 International License.
+ *
+ * You are free to:
+ *  - Use, copy, modify, and distribute this file
+ *
+ * Under the following conditions:
+ *  - Attribution: You must give appropriate credit to the author.
+ *  - NonCommercial: You may not use this work for commercial purposes.
+ *
+ * License details:
+ * https://creativecommons.org/licenses/by-nc/4.0/
+ */
+
+package com.bitlabbr.minhadespensa.uisystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val PrimaryLight = Color(0xFF6200EE)
+val SecondaryLight = Color(0xFF03DAC6)
+val BackgroundLight = Color(0xFFFFFFFF)
+
+val PrimaryDark = Color(0xFFBB86FC)
+val SecondaryDark = Color(0xFF03DAC6)
+val BackgroundDark = Color(0xFF121212)
