+++ uisystem/src/commonMain/kotlin/com/bitlabbr/minhadespensa/uisystem/theme/Type.kt
@@ -1,0 +1,22 @@
+/*
+ * Copyright (c) 2026 Willian Santos
+ *
+ * This file is part of an open-source project licensed under the
+ * Creative Commons Attribution-NonCommercial 4.0 International License.
+ *
+ * You are free to:
+ *  - Use, copy, modify, and distribute this file
+ *
+ * Under the following conditions:
+ *  - Attribution: You must give appropriate credit to the author.
+ *  - NonCommercial: You may not use this work for commercial purposes.
+ *
+ * License details:
+ * https://creativecommons.org/licenses/by-nc/4.0/
+ */
+
+package com.bitlabbr.minhadespensa.uisystem.theme
+
+import androidx.compose.material3.Typography
+
+val AppTypography = Typography()
